Move Bloom filter into separate file
