+++ worker/util/kv-manager.ts
@@ -1,5 +1,4 @@
-import PromiseGatherer from '../promise_gather';
-import { getWASMModule } from '../wasm';
+import { BloomFilter } from './bloom-filter';
 
 declare const KV_CACHE_NAMESPACE: KVNamespace;
 
@@ -9,74 +8,8 @@ export enum KVExpiration {
     NONE
 }
 
-const ALLOCATED_ROWS = 10;
-const K_HASHES = 3;
-
-class BloomFilter {
-    private static seed = 8149214274; // Randomly chosen, nothing special
-    private static m = ALLOCATED_ROWS;
-    private static k = K_HASHES;
-
-    public static async add(element: string): Promise<void> {
-        if (await KVDirect.get('bloom:0') === null) {
-            await this.allocate();
-        }
-
-        const gatherer = new PromiseGatherer();
-
-        const indexes = await this.getIndexes(element);
-        for (const index of indexes) {
-            gatherer.push(KVDirect.put('bloom:' + index, '1', KVExpiration.TIMED));
-        }
-
-        return gatherer.all();
-    }
-
-    public static async has(element: string): Promise<boolean> {
-        if (await KVDirect.get('bloom:0') === null) {
-            await this.allocate();
-            return false;
-        }
-
-        const indexes = await this.getIndexes(element);
-        for (const index of indexes) {
-            if (await KVDirect.get('bloom:' + index) === '0') {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    // See: https://willwhim.wpengine.com/2011/09/03/producing-n-hash-functions-by-hashing-only-once/
-    private static async getIndexes(element: string): Promise<number[]> {
-        const [a, b] = await this.doubleHash(element);
-        let indexes = [];
-        for (let i = 0; i < this.k; i++) {
-            indexes.push((a + b * i) % this.m);
-        }
-        return indexes;
-    }
-
-    private static async doubleHash(value: string): Promise<number[]> {
-        const wasm = await getWASMModule();
-        return [ Number(wasm.xxhash(value, this.seed)), Number(wasm.xxhash(value, this.seed + 1)) ];
-    }
-
-    private static async allocate(): Promise<void> {
-        const gatherer = new PromiseGatherer();
-
-        for (let i = 0; i < this.m; i++) {
-            gatherer.push(
-                KVDirect.put('bloom:' + i, '0', KVExpiration.TIMED)
-            );
-        }
-
-        return gatherer.all();
-    }
-}
-
 export class KVManager {
-    public static async get(key: string): Promise<string | null> {
+    static async get(key: string): Promise<string | null> {
         const seen_before = await BloomFilter.has(key);
         if (!seen_before) {
             return null;
@@ -85,7 +18,7 @@ export class KVManager {
         return KVDirect.get(key);
     }
 
-    public static async put(key: string, value: string): Promise<void> {
+    static async put(key: string, value: string): Promise<void> {
         const seen_before = await BloomFilter.has(key);
         if (seen_before && await KVDirect.get(key) === null) {
             return KVDirect.put(key, value, KVExpiration.PERIODIC);
@@ -95,14 +28,14 @@ export class KVManager {
     }
 }
 
-class KVDirect {
+export class KVDirect {
     private static namespace = KV_CACHE_NAMESPACE;
 
-    public static async get(key: string): Promise<string|null> {
+    static async get(key: string): Promise<string|null> {
         return this.namespace.get(key);
     }
 
-    public static async put(key: string, value: string, expiration = KVExpiration.NONE): Promise<void> {
+    static async put(key: string, value: string, expiration = KVExpiration.NONE): Promise<void> {
         switch (expiration) {
             case KVExpiration.PERIODIC:
                 return this.namespace.put(key, value, { expirationTtl: 86400 });
